api_management: splitting the ID parsers
