+++ azurerm/internal/services/apimanagement/parse/diagnostic_id_test.go
@@ -2,89 +2,6 @@ package parse
 
 import "testing"
 
-func TestApiManagementLoggerID(t *testing.T) {
-	testData := []struct {
-		Name     string
-		Input    string
-		Expected *ApiManagementLoggerId
-	}{
-		{
-			Name:     "Empty",
-			Input:    "",
-			Expected: nil,
-		},
-		{
-			Name:     "No Resource Groups Segment",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000",
-			Expected: nil,
-		},
-		{
-			Name:     "No Resource Groups Value",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/",
-			Expected: nil,
-		},
-		{
-			Name:     "Resource Group ID",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/",
-			Expected: nil,
-		},
-		{
-			Name:     "Missing Service Name",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.ApiManagement/service/",
-			Expected: nil,
-		},
-		{
-			Name:     "Missing Logger",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.ApiManagement/service/service1",
-			Expected: nil,
-		},
-		{
-			Name:     "Missing Logger Value",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.ApiManagement/service/service1/loggers",
-			Expected: nil,
-		},
-		{
-			Name:  "Logger ID",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.ApiManagement/service/service1/loggers/logger1",
-			Expected: &ApiManagementLoggerId{
-				Name:          "logger1",
-				ServiceName:   "service1",
-				ResourceGroup: "resGroup1",
-			},
-		},
-		{
-			Name:     "Wrong Casing",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.ApiManagement/service/service1/Loggers/logger1",
-			Expected: nil,
-		},
-	}
-
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Name)
-
-		actual, err := ApiManagementLoggerID(v.Input)
-		if err != nil {
-			if v.Expected == nil {
-				continue
-			}
-
-			t.Fatalf("Expected a value but got an error: %s", err)
-		}
-
-		if actual.Name != v.Expected.Name {
-			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
-		}
-
-		if actual.ServiceName != v.Expected.ServiceName {
-			t.Fatalf("Expected %q but got %q for Service Name", v.Expected.Name, actual.Name)
-		}
-
-		if actual.ResourceGroup != v.Expected.ResourceGroup {
-			t.Fatalf("Expected %q but got %q for Resource Group", v.Expected.ResourceGroup, actual.ResourceGroup)
-		}
-	}
-}
-
 func TestApiManagementDiagnosticID(t *testing.T) {
 	testData := []struct {
 		Name     string
